requirements.txt
python-dotenv
SQLAlchemy~=2.0.41
python-telegram-bot~=22.2